+++ desktop/cmp/dash/DashModel.ts
@@ -5,7 +5,7 @@
  * Copyright © 2023 Extremely Heavy Industries Inc.
  */
 import {bindable, makeObservable, observable} from '@xh/hoist/mobx';
-import {HoistModel, managed, PersistenceProvider} from '@xh/hoist/core';
+import {HoistModel, managed, PersistenceProvider, RefreshContextModel} from '@xh/hoist/core';
 
 /**
  * Base Model for {@link DashCanvasModel} and {@link DashContainerModel}.
@@ -17,6 +17,8 @@ export abstract class DashModel<VSPEC, VSTATE, VMODEL> extends HoistModel {
     viewSpecs: VSPEC[] = [];
     @observable.ref state: VSTATE[];
     @managed @observable.ref viewModels: VMODEL[] = [];
+
+    @managed refreshContextModel: RefreshContextModel;
 
     //-----------------------------
     // Settable State
@@ -43,5 +45,7 @@ export abstract class DashModel<VSPEC, VSTATE, VMODEL> extends HoistModel {
     constructor() {
         super();
         makeObservable(this);
+
+        this.refreshContextModel = new RefreshContextModel();
     }
 }